+++ CHANGELOG.md
@@ -35,6 +35,7 @@ This release adds support to OneView Rest API version 500 for the hardware varia
    - Switches
    - Unmanaged Devices
    - Uplink Set
+   - User
 
 #### Breaking changes
 1. In the Switch resource, the `statistics` method was fixed, a non-existent endpoint was removed `/rest/switches/statistics/{portName}/subport/{subportNumber}`. Now the method only get statistics of a switch or a specific port.